Inject strings from meson.build
In this commit, a mod called config is dynamically generated using a
template at src/config.rs.in, which is populated by Meson when compiling
the code.

This concept will be used next in order to add support for resource
files and translations, but at the moment I make this simple commit to
show the most basic way for achieving this.

So now I can un-hardcode the application ID and stuff like the version
name and just use whatever Meson is providing on compile time.