Use Meson to compile gresource files
The next step in the tooling adventure is to stop using raw .ui files
and start bundling all the resource files into a .gresource file. Note
that I am not a fan of having the resources in a sidecar file when I
could just embed them into the binary, but I guess that this is just a
matter of taste.

Briefly, I create a file called cartero.gresource.xml which describes
the layout for a cartero.gresource file. The XML file describes the list
of files that will be packed, which are the .ui files part of the user
interface. Note that this commit does not touch the ./compile.sh script
yet: Blueprint will be provided in a separate commit.

I use this occasion to cleanup and rename some files. In particular,
since many other resources will be in the data/ directory, it is
probably wise to move the .ui files into data/ui. Also, rename the
prototype.blp file into something like main_window.ui, because it is not
a prototype anymore.

Then configure Meson to build the cartero.gresource file, and expose it
via another build variable into the config.rs.in file.

The key thing here is to replace the Rust template macros so that
instead of hardcoding an .ui file, it uses the resource identifier, and
updating the main.rs file so that the gresource file is loaded on
application startup.