Use Meson to compile Blueprint files
The next step is to follow the install instructions located at
https://jwestman.pages.gitlab.gnome.org/blueprint-compiler/setup.html in
order to make compiling the .blp files into .ui files part of the build
process.

I add a wrap for blueprint-compiler, which means that while Meson will
try to use the system blueprint-compiler if provided, it will download
it from upstream otherwise. Then during build time it will translate the
.blp into .ui and leave them in the build directory for the gresource
step to pick it up.