Enforce formatting rules for Blueprint
Since Blueprint 0.12 has a formatter, configure the hook so that it also
enforces formatting rules in .blp files prior to allowing the developer
to commit.

Apparently I found some odd behaviour in the formatter that I don't know
yet if it is a bug or not in regards to whitespace in multiline
comments. I have reported it in the issue tracker for
blueprint-compiler, but in the mean time, I will change the copyright
notice comments in .blp files so that they use a different style that
doesn't cause the formatter to change whitespace on them.