Add internationalization for this project
Finally, install support for gettext so that Meson can be used to
update and compile translation files (.mo). Install the gettext-rs crate
in order to use these translations from the Rust application.

Updates the .blp files to provide some translations to begin with, and
add instructions in regards to how to update the translations in the
README.md file.