Bind a Header object with the HeaderRow widget
(If you watched my development session earlier today, I am a whole clown,
because it turns out I wasn't updating the actual gresource that enters
when running the application via cargo run. To prevent this from
happening in the future, I've added a script to build-aux/cargo-build.sh
that will heavy lift all the cargo build for you.)

This commit is the first attempt at using property binds at the
templates. The end goal is to have the row widgets backed by different
CarteroHeader objects instead of having a mess of GtkEntry and
GtkCheckButton.

In this commit, the RowHeader component is refactored so that it has a
backing `header` property, and then the template is bound so that the
value is backed by the header entity itself.

At the moment, this backing is not bidirectional, which means that the
values are not stored automatically in the model itself. That's a future
task.