Fill the ListView dynamically (In progress)
Finally I am using a GtkListView and I'm using factories and models and
that stuff to show the list of headers.

Note that this is a WIP. At the moment, this commit is able to present a
list of headers stored in the ListModel using the given ListFactory. Now
I only need to be able to update the ListModel when I change the widget
content. And of course to add new rows. But hey, you can press Send and
see the list of headers in your console.