Implement the request header list
This commit adds a widget called RequestHeaderPane whose purpose is to
allow the user to add, modify and remove headers based on the work done
before.

This is a prototype implementation that I am sure it has bugs, but it is
too late to test this widget. Anyway, the following things should be
supported:

* You should be able to add new headers by pressing the Add button.
* You should be able to set the content of the header, both name and
  value.
* You should be able to deactivate headers using the checkbox.
* You should be able to delete headers by pressing the Trash button.

Currently, pressing the Send button will list every header in the
console, but in the future this will be changed with the actual code
that adds the headers to the HTTP request being sent.

Known issues:

* Focus does not work properly, so tabbing through the panel is
  difficult right now.
* The trash button could be replaced with something more friendly, such
  as a [...] button that opens a menu where one of the options is
  Delete.
* Headers may be swalowed at any time or not being deleted.