Just default to UTF-8 encoding
Yesterday I had issues with String::from_utf8() because it fails if the
string is encoded as ISO-8859-1 and uses characters from the upper 128
character set that cannot be converted back to UTF-8.

Rather than just noticing that there is a String::from_utf8_lossy()
method that will convert invalid characters to the <?> character, I
added a fallback for ISO-8859-1. However, this doesn't scale and it is
probably just better to use the lossy variant.