Cosmetic changes to the application
The FlowBox was responsive, which is one of the reasons why it was being
used. However, unlike a Paned, it doesn't allow to change the size of
each pane.

I've replaced it back with a Paned, however because I am using
libadwaita now, I can use Adwaita breakpoints to make the Paned
responsive, and turn it vertical when there is not enough horizontal
space.

I think that the responseive flowbox was a great enhancement over a
standard Gtk.Paned, but allowing the user to change the proportions of
each pane seems promising, and this is one of the places where having
Adwaita breakpoints makes things easier for everyone.

Some other changes made to the application:

* Some control groups such as the address bar or the KeyValueItem box
  have been linked to be presented as a single slab.
* Some borders have been polished out to make a better use of the space.