Asynchronous requests
The requests shall be done asynchronously to avoid making a slow request
from freezing the entire UI. Take <https://httpbin.org/delay/5> as an
example.

After this commit, isahc uses async requests, which has forced me to
also make some methods async-await. I had to remove the TryFrom to cast
a isahc Response into a local response object, since I had to make that
method async as well.

A spinner is added to the response panel so that it shows that a request
is in action, specially given that during a request, the panel is
locked.