Serialize active/secret metadata
After this commit, if a header or a variable has a different value for
the active checkbox or the secret checkbox, the value will be stored
using a full entry with the metadata.

This allows to keep it short (header = "value") if the user hasn't
touched anything, but to have metadata if there are more things to save.
Similar to how Cargo.toml uses strings or objects depending on the
value.

Also works for reading and I have tested that it is safe with existing
files.