Change prefix location when running as a Mac app
macOS regular applications use a different directory structure than UNIX
applications. The executable lives in Cartero.app/Contents/MacOS, and
resource files (such as the datadir) should live in
Cartero.app/Resources/share, Cartero.app/Resources/lib, and so on.

This commit will add a codepath specific for macOS that allows using the
Resources directory as the real rootdir if the directory exists. After
this commit, therefore, the application should also run if Cartero is
properly packaged as a macOS application, since the resource files are
expected to be found ni the Resources dir.

Note that I haven't just changed the URL. I'm testing that Resources
exist, and falling back to the standard ../ root directory like it
happens in Windows and GNU/Linux, in order to make the
build-aux/cargo-build.sh script also work on macOS for quick development
purposes.