FileDialogs stopped being native on Windows
It turns out that if you add a MIME filter to a Gtk.FileFilter, then the
Gtk.FileDialog on Windows stops being native and is the GTK-based one
again. So in this commit I avoid running that code on Windows, to make
sure that the dialog is the Windows one.

Also, it seems that having the file format and the extension is
something that is different between operating systems, so the string to
place for the FileFilter name is now also based on the operating system.

Extra minor things:

* Allow the user to override the GSK_RENDERER variable if wished (mainly
  will be used to test the GL renderer while keeping Cairo the default
  renderer on Windows).
* msys-build.sh now requires you to specify the development or the
  default version to compile.