Fix loading locales for third party libraries
It turns out that dependencies have their own locale files, and this may
explain why the load and save dialog on Microsoft Windows always used
the English messages for the Cancel button, or why the About Dialog was
not translated only on Microsoft Windows.

This commit patches the msys-build.sh script so that third party locales
are copied when creating a binary distribution for Microsoft Windows.

This commit also patches main.rs so that when running the application on
Windows, libadwaita is rebound again to the proper directory, for
reasons stated in the comment I've added on top of the block.