Fix word wraping on long lines without spaces
As described in https://github.com/danirod/cartero/issues/67, by default
if a long line has no spaces, a Gtk TextView cannot find a break point
for the word wrapping. The solution is to use WordChar, which uses
characters as a fallback if a line has no spaces and it cannot
physically break by word.