Allow to open more than one file at the same time
It is already possible to open multiple files using the file manager,
but it is currently not possible to pick multiple files from the Open
dialog.

In this commit I've patched the open_file() method to have a separate
open_files() instead that opens multiple files. I am keeping the
open_file() around in case I still need it in the future. The win.open
action now will invoke open_files() and open every file returned by the
function call.