Split with Serializable interface
There is a new interface called Serializable. It allows a widget to
convert its data into a TOML string and to read data from a TOML string
by using the from_toml() and to_toml() virtual methods.

The EndpointPane has been modified so that it implements this interface
and knows how to persist its data through the RequestFile struct.

The ItemPane is finally using interfaces to communicate. It knows how to
load and save data from Serializable objects. As long as the child is a
Serializable, the load_pane() and save_pane() methods can be used to
load data from a gio::File and to save data into a gio::File by using
the Serializable virtual methods to get or assign the TOML
representation.