Check special datafiles
These are the GLib schema file, the desktop file and the AppStream
distribution file. These are the files that usually are tested when
calling `meson test`. Unfortunately I have to type the commands manually
because `meson test` in the CI environment is picking also the test
files for blueprint-compiler for some reason and they fail in my
environment.