Move URL checking to BoundRequest::try_from
After this commit, the tests to assert that an URL has a valid
protocol (as in, it's not an unsupported protocol), or a protocol at
all, are moved to the BoundRequest initialization rather than keeping
them in the EndpointPane.

The behaviour of prepending http:// to the typed URL is still in
place, and as can be seen in the diff, the
RequestPreconditionError::MissingProtocol is a special case that
allows to do this and repeat the request.

Also, the URL normalization (as in, removing leading and trailing
spaces) is also done here. The BoundRequest will have the URL set to
whatever is being normalized by the url crate.

Since more errors have been moved from CarteroError to
RequestPreconditionError, a cleanup has been made in CarteroError to
remove the keys that are not in use anymore.