Compact Field declarations
This change is related to the one in the previous commit of this set.
Due to its nature, I've made a separate commit for it to make it easy to
rollback if it continuously fails.

I've squashed a few file data structures, I've removed From<FieldValue>
implementation for KeyValue and replaced it instead with
FieldValue::to_key_value(), since the desired name of the field has to
be given anyway.