Generalize serialized key-value containers
FieldValue (previously known as KeyValuedValue) has to be different for
query params, since we only persist inactive query parameters (the
active query params are kept in the URL).

In the future, a separate structure will exist to persist inactive query
params. There is no need to persist the active query params because they
can be present in the URL.

However, it seems that currently inactive query params are ignored at
all within the user interface, so that will need to be fixed first.