Don't lose disabled query params on URL change
After applying this commit, updating the URL entry will not delete the
query params that are disabled in the parameters table. This is done
by saving the disabled parameters and appending them to the table
after each update.

I haven't checked the performance implications when there are a lot of parameters.