Merge CSD/non-CSD declarations of CarteroWindow
After learning how to use cfg_attr(), this commit merges the
CarteroWindow struct that is used when CSD is active, and the one used
when CSD is not active.

They used to be separate structures because I need to pass a differente
template resource depending on whether CSD is enabled or not, but I've
learned how to use cfg_attr(), which allows to add an attribute only if
a cfg() predicate is true, so I can merge them now as the commit shows
very well.