Remove any existing tab bindings
Since I am about to replace this with SignalGroup and BindingGroup, this
commit removes bindings for the window actions (win.save, win.save-as,
win.request and win.close), as well as setting the headerbar title and
subtitle when the application runs in CSD mode.

This will allow me to work on the new implementation from scratch.