Make UI fonts larger on macOS (#149)
This commit will add custom CSS rules to style.css that make the
application have a larger font size when running on macOS. This is done
by adding conditional classes and hacking the DPI of the window in a way
that hope doesn't bite me in the future. I have tested in HDPI mode and
fortunately macOS does not actually expose a way to customize the DPI.

Note that there are hacks inside the hack.

- I'm only patching the widgets that I see are in use by the
  application.
- Patching the popovers is tricky because if you scale it too much, it
  adds ellipsis, so I keep a smaller font size for these components.