Connect tab events via signals
After this commit, things such as the window title and subtitle, and
whether the actions linked to an opened pane, such as save, save-as and
close, will be bound to the tab bar via regular binds.

Using a BindingGroup I am able to connect a single TabPage to multiple
target properties, and even I can use a transformation function to apply
some custom mapping in the process to the window title for non-CSD
compilations.

It seems that the BindingGroup is not called when the source is set to
NULL, so there is additional logic to reset the title when the last
window is closed.