Refine application icon (#153)
# Motivation

The current Cartero icon has too many details that are lost when
presented in low resolution (such as in lists or as a small menu icon).
Refine the existing icon removing some parts to make it more. Also,
create a square version of the icon to fix #131.

# New icon

![An envelope with red and blue stripes and a pair of arrows in the
middle](https://github.com/user-attachments/assets/7438ebd9-a57d-4684-8152-9bbeeb3d0a79)

Changes:

* There are less stripes in the border and the stripes are bigger, to
make them more noticeable.
* The arrow colors match the colors of the stripes.
* The stamp has been removed because it is not visible in low
resolutions.
* The icon side colors have been updated.

# Rebranding

The icon is currently in use at least in the following sites:

* The project website at https://cartero.danirod.es
* The user manual at https://cartero.danirod.es/docs
* The Snapcraft page at https://snapcraft.io/cartero
* The Flathub page at https://flathub.org/apps/es.danirod.Cartero
* The social media profile at https://fosstodon.org/@cartero

The icons and avatars in use at this locations will not be updated until
this icon is added to the next standard release (0.1.6 or, most probably
at this point, 0.2.0). However, this is a necessary step to do.

## Flathub review status

Back when I submitted the original application to Flathub, the previous
icon got a negative review of "In line with contemporary styles". It is
NOT a priority to pass this check, as this is a very subjective topic
and I don't think Flathub is in a position to jugde application icons.