Add Generative AI policy (#154)
## Motivation

Ban AI contributions from the code before it's too late, by adopting a
similar guidelines to the ones found in
https://discourse.gnome.org/t/loupe-no-longer-allows-generative-ai-contributions/27327.

## Rationale

As described in the original thread, using code generated by an LLM
could pose a legal threat if the code regurgitated by the parrot is
based on something with a license that is not compatible with the one in
this project. Besides, out of context code generated by an LLM can have
low quality standards that make maintenance difficult.

(The following is an off-the-record statement by me, that is not in the
original thread and it's not representative of the whole GTK and GNOME
project.) This is even more truthful in the context of GTK and gtk-rs,
which are libraries whose documentation is more scarce, so the knowledge
gained by LLMs is not big enough. Also, given that the libraries and the
SDK is in evolution, the code generated by the LLM could be outdated
already.