Separate RequestBuildError
If the conversion from BoundRequest to the native Request object fails,
a new error kind RequestBuildError will be thrown.

To be fair, I don't think this error is going to be used frequently.
Many checks are already done by BoundRequest, catching things such as
invalid port numbers. However, validating that a header has a valid name
will be caught as a RequestBuildError.