What if we collectively made as if ItemPane never existed?
In a future change, a new ItemPane will be added using interfaces and
traits, similar to the work done previously for
https://github.com/danirod/cartero/pull/115 and
https://github.com/danirod/cartero/pull/117. However, until that
happens, let's just remove ItemPane and make the CarteroWindow TabView
add straight EndpointPane entities.

I've moved the two properties of ItemPane, file and dirty, and I've
updated every other reference to the downcasts made to ItemPane so
that they now get an EndpointPane instead.