Use a local action to trigger perform_endpoint
This commit will replace win.request as the action that actually
triggers a call to EndpointPane::perform_request() to start the
request. Instead, each EndpointPane has now its own "request" action
that calls the function locally.

"win.request" is kept, because global keybindings can only be
associated with window-level actions in GTK-RS currently, but this
action will only delegate a call to the local action if a tab is
visible.

Because now I'm using local actions, I'm also replacing the previously
existing callback that updates the activeness of the Request button in
the EndpointPane, and I'm using bindings instead, to set whether the
action can be used depending on whether the current URL is set or empty.