Define read-only mode for EndpointPane
This commit adds a couple of properties to EndpointPane and all the
associated widgets to introduce the "read-only" property. When
enabled, the inputs will not be editable or actionable.

The main difference between setting EndpointPane:read-only and using
EndpointPane:sensitive is that making the pane not sensitive at all
will also disable the notebook tabs and any additional buttons that
may be added in the future (such as the cancel button). By using
read-only I can cherry pick what controls I actually want disabled,
letting the rest of the pane sensitive by default.