Restore non-blocking behaviour of EndpointPane
It seems that you either catch panics or let the EndpointPane request
performing behave async. Maybe I need to add the catch_unwind()
somewhere else in the code, but for the moment let me treat this
commit as a checkpoint so that I can restore the intended behaviour,
because I want the EndpointPane (and the CarteroWindow at all) to stay
sensitive and continue responding to the system while a request is in progress.