Early implementation of request cancelling
This commit adds a button that stops requests. To achieve this, the
JoinHandle of the thread spawned via glib::spawn_future_local() is
stored as a reference that the new "endpoint.cancel" action clears if
set.

Drawbacks: currently it seems that my implementation does not actually
close the TCP socket when the request is cancelled. According to
isahc's docs, this should be possible. However, it's not very clear
when that happens. I've tried to make my custom wrapper and implement
Drop on it, but it doesn't work.

Maybe this is a signal that I should drop isahc and switch to hyper-rs.