Add a shortcut to focus the window URL
As proposed in https://github.com/danirod/cartero/issues/156, this
commit will add a shortcut to the EndpointPane class that will cause
the request URL field to be focused if the Ctrl-L (or Cmd-L key in
macOS) combination is pressed.

The interesting thing about this commit is that I did not know that
you were supposed to add shortcuts to the Blueprints in GTK 4, but now
I know. These are more refactors for the future.