Initial approach to font zooming
There is no such thing as "font scale", but it is possible to multiply
the font-size by some ratio in order to make it bigger or smaller. This
is what this commit does by linking a multiplier to the current zoom
level, which is a number that can be set using the zoom-in, zoom-out and
zoom-reset actions defined in CodeView locally.