Create POC for version checker
Do not merge. It uses a hardcoded URL that points to localhost. Also
doesn't do much apart of printing to the console the response of the
server if there is a new version.

And still, the contents of this commit would allow to parse the server
response for the GitHub latest release endpoint and detect whether the
latest version in GitHub is newer than the one currently running.