Design and add authorization pane to UI
This commit introduces the visual part of this PR by adding a specific
pane for authorization.

Unlike the request payload pane or the export pane, I am adding here all
the handling in a single object. The point I am trying to proove here is
what is more maintainable: to have a multiple entity system, or to just
add all the code into a single file. This will be a race to see which
system is the one who breaks first.

Either way, the widget is very simple and has a couple of inputs that
are visible or hidden depending on the value of the selection method
picked in the combo box. There are methods to set the UI state based on
the RequestAuthorization struct, and to take a RequestAuthorization
struct from the UI state.

This is currently being added to EndpointData and it is being verified
that loading a file will import the current authorization settings, and
saving a file will also export the current authorization settings.