Define entity for request authorization
There are different kinds of request authorization protocols and each
one may supply their own set of parameters and behaviours. Therefore, it
makes sense that each kind of supported authorization protocol is a
separate branch in an enum called RequestAuthorization.

Currently, as a proof of concept, RequestAuthorization will only support
bearer tokens and basic authentication. Note that this commit will only
provide the data structures and the serialization to persist and load
data from files, but will not actually encode the headers now.