Check for license compatibility
Because Cartero is published under GPL-3.0, I need to make sure that
dependencies are compatible with mine, because this is Rust so
everything will be tangled into one final binary, and the rules of
GPL-3.0 apply to the code I import.

The list of licenses that are valid include GPL-3.0 and LGPL-2.1, of
course. Also include MIT, BSD and Apache-2.0, because there is
confirmation that these are compatible with the GPL-3.0, and then
licenses that have been clarified compatibility like Unicode-3.0 and
CC0-1.0.