Update project dependencies
Runs cargo update and cargo upgrade to put dependencies up to date. Runs
cargo t and cargo r to make sure that the project still runs and works
as expected.

Only fatalities are that srtemplate::SrTemplateError is gone because
it's know now as just srtemplate::Error, so I update the import in my
conversions.