Update project dependencies (#168)
I don't remember when was the last time that I did this.

## Notable things

* First, traitobject has finally been upgraded to 0.1.1, this is going
to remove the warning about "future compatibility" received every time
Cartero is compiled.
* In general, traitobject has to go, because it is a transitive
dependency of formdata, which is a crate that has been officially
archived. I have to look for alternatives to generate the
multipart/form-data body of a request, and then remove formdata, because
of the amount of transitive dependencies it introduces.
* srtemplate changed the name of their error type from SrTemplateError
to just Error. The code has been updated to use the new name.

## License check

This PR also introduces `cargo deny` as a tool to inspect for issues
with dependencies. In the meantime I only check the license
compatibility. If at some point a dependency that is not compatible with
GPL-3.0 is added, it will complain. This is important because Cartero is
statically linked due to how Rust works.