Remove i18n wrapper and prepare for gettext 0.24
GNU gettext 0.24 is out and one of the new features is a dedicated
parser for Rust [1], which means that .rs files can now be parsed for
strings without having to fallback to the C parser.

The updated manual contains recommendations for Rust projects [2], such
as using the gettext-rs crate rather than gettext or xtr (as we do), or
using formatx for formatted strings.

However, I need to make sure that the code is still compatible with
older versions of GNU gettext, because breaking compatibility may break
some of the platforms where Cartero currently compiles, like Ubuntu
24.04 (where gettext 0.21 is provided).

Therefore, I'm going to use these recommendations already by removing
the i18n wrapper that I previouslly pulled from another GNOME project
written in Rust [3], and using instead the vanilla gettext-rs functions
for most cases.

Code that used to call i18n_f to pass parameters is now using formatx.
Previously this was done like:

    i18n_f("Method: {}", &[&method])

Now it's done like this:

    formatx!(gettext("Method: {}"), method).unwrap()

Notable changes:

* i18n_f never fails, formatx!() returns a Result that has to be
  unwrapped to extract the string.
* Since the first parameter of i18n_f was treated as a gettext function
  (po/meson.build), the string does not change. Our cartero.pot file
  contains rust style bracket placeholders ("Method: {}"). In fact, this
  is also the standard way for Rust formatted strings in GNU
  gettext 0.24 [4].

Code that previously passed a string for singular and a string for
plural is now using ngettext() rather than gettext(), which behaves the
same way.

[1]: https://users.rust-lang.org/t/gnu-gettext-now-supports-rust/126224
[2]: https://www.gnu.org/software/gettext/manual/html_node/Rust.html
[3]: https://gitlab.gnome.org/World/Solanum/-/blob/main/src/i18n.rs
[4]: https://www.gnu.org/software/gettext/manual/html_node/rust_002dformat.html