Add authorization pane (#167)
Implements #60, adding a new pane to set the Authorization details using
a form rather than having to manually set the Authorization header
manually in the headers pane. Currently supported authorization methods
are Basic Auth and Bearer Tokens, but I'm very interested in adding more
methods in the future, like JWT, OAuth 2.0 or AWS4.

Each authorization method can later modify the request when it's being
bound for a request. This makes possible to also assign variables to the
values of these headers.