Update copyright years in about window and source code headers (#172)
I never bothered to update the copyright year shown in the About dialog.
This PR will fix that.

Additionally this commit will update every source code header that is
still outdated or missing. Hopefully something can be added before
December to update all these files for the next year, or at least
something could be added as a pre-commit hook to check that added files
have an SPDX identifier and the years.