Fix bugs after updating to GNU gettext 0.24
So, gettext has support for Rust, but it seems that it does not support
parsing strings that are inside closures with callbacks, such as
glib::clone!(). To keep those strings, I have to rewrite the code in a
way that does not make use of strings in closures, such as calling a
separate function inside the callback.