Fix bugs after updating to GNU gettext 0.24 (#177)
* gettext is now more strict when validating `translate="no"` nodes in
XML files like AppData. Add `translatable="no"` to fix it, but keep the
old `translate="no"` in case there are tools using it.
* The new Rust parser is unable to parse strings inside macros that make
use of closures. To keep it simple, move everything that can be moved
outside of the glib::clone!() or glib::closure!() macros that is
currently using translatable strings inside.