Disable updater unless requested during compile
The updater is optional and is disabled by default. The primary reason
behind this is to make things easier for environments where the package
manager already manages updates in behalf of the user (such as Flatpak,
Snapcraft, AUR, dpkg...).

The updater is meant to be enabled in the AppImage executable, and the
Windows and macOS versions of the app, because these are standalone and
downloaded from the project website. However, many software repositories
already assume that updates will be disabled because the package manager
will handle that.

To enable the updater, the `-Dapp-updater=enabled` flag will have to be
passed during compilation. If not given, the default (disabled) will not
include the updates module during compilation, and will not display the
button in the settings page.